Fixing errors to the cultural page and footer for all the web pages.
Fix: fixing the footer background being cut out. Fixed this to all the web pages' footer.

Fix: fixed the space and size of images on the cultural page because image size and spacing doens't workk properly on the desktop.

css: on festival and cultural image,  changing image size from px to %. For footer of all the pages, changed the padding px to % to fix the cut out.