Finish Index and About page
Feat: I made the navigation sticky and added the scroll js to change the colour of the nav when scrolled down to specific section.

Fixed: I fixed and added the link to my social media icons and resources.

I changed the aesthetic of the social meida icons.

Changed the dummy text to text with context on topic section.

Home page and about page is done!